Move LifecycleHandle to its own module.
