+++ overwatch/src/services/lifecycle/mod.rs
@@ -1,0 +1,7 @@
+pub mod handle;
+pub mod message;
+pub mod notifier;
+
+pub use handle::LifecycleHandle;
+pub use message::LifecycleMessage;
+pub use notifier::LifecycleNotifier;
